--- next.config.js
@@ -1,8 +1,0 @@
-// next.config.js
-module.exports = {
-    images: {
-      domains: ['encrypted-tbn0.gstatic.com', 'images.unsplash.com', 'cdn.example.com'], // Add your image sources here
-    },
-    reactStrictMode: true,
-  };
-
